refactor auto functionalize test in their own file
[ghstack-poisoned]
